Produce a backwards-compatible Java 9 module
This adds the necessary definitions to produce a jar file that will
behave as a real module on Java 9, but will also execute correctly
on Java 6 and above.

This adds a maven-enforcer-plugin rule to require JDK 9 to build
the code (and to give a friendly error message if the user attempts
to build on a JDK older than 9). It adds two compiler executions:
One that compiles all of the code as Java 9 bytecode in order to
get accurate error messages with regards to module problems (such
as missing "requires" directives, using APIs that are not in Java 6,
etc), and another execution to recompile all of the code except for
the module descriptor as JDK 6 bytecode.

It also adds a temporary dependency on Bnd 3.5.0 in order to work
around FELIX-5698.

See: https://issues.apache.org/jira/browse/FELIX-5698
Affects: #29