Deprecate the UUID scheme (#47) (#77)
* Remove usage of IDs in BroadphaseKey; use reference equality instead

* Remove Fixture and AbstractCollidable override of equals().

The previous implementation tested the equality of the object IDs. Since
there's a 1-1 mapping to IDs and those objects the java operator == does
the same job. The overriden methods is removed since the default
implementation of equals is the operator ==

* Remove AbstractShape hashCode() cause it does not adhere to the contract

hashCode should not be implemented when equals is not implemented.
Here since equality does not make sense other than the default equals(),
hashCode() should be the default implementation as well.

* [forgot to] Remove Fixture and AbstractCollidable override of hashCode()

* Remove direct use of IDs in ContactConstraintId. New getters fields

* Update javadoc in BroadphaseKey + slight toString() change

* Deprecate public getId() API, for removal

* Remove unnecessary use of getId() in BroadphaseItem

* Do not use IDs to compare SapProxy objects

* Remove getID() unit tests

* Small javadoc fix